+++ src/app/services/global-settings/global-settings.service.ts
@@ -88,8 +88,8 @@ export class GlobalSettingsService {
     this.logger.debug('[GLOBAL-SET] setVariableFromStorage :::::::: SET VARIABLE ---------->', Object.keys(globals));
     for (const key of Object.keys(globals)) {
         const val = this.appStorageService.getItem(key);
-        // this.logger.debug('[GLOBAL-SET] setVariableFromStorage SET globals KEY ---------->', key);
-        // this.logger.debug('[GLOBAL-SET] setVariableFromStorage SET globals VAL ---------->', val);
+        this.logger.debug('[GLOBAL-SET] setVariableFromStorage SET globals KEY ---------->', key);
+        this.logger.debug('[GLOBAL-SET] setVariableFromStorage SET globals VAL ---------->', val);
         if (val && val !== null) {
             // globals.setParameter(key, val);
             globals[key] = stringToBoolean(val);
@@ -111,8 +111,10 @@ export class GlobalSettingsService {
     }
 
     TEMP = getParameterByName(windowContext, 'tiledesk_supportMode');
+    console.log('TEMP supportMode', TEMP);
     if (TEMP) {
         globals.supportMode = stringToBoolean(TEMP);
+        this.appStorageService.setItem('supportMode', String(globals.supportMode))
     }
 
     TEMP = getParameterByName(windowContext, 'tiledesk_lang');
@@ -138,6 +140,7 @@ export class GlobalSettingsService {
     TEMP = getParameterByName(windowContext, 'tiledesk_projectID');
     if (TEMP) {
         globals.projectID = TEMP;
+        this.appStorageService.setItem('projectID', TEMP)
     }
 
     TEMP = getParameterByName(windowContext, 'tiledesk_logOut');